Adjust default touch calibration for Core2
p->touch(t) call invokes Panel_Device::touchCalibrate and
results in setting up the affine matrix to map the touch
geometry into the panel one. This is not preferable as the
touch geometry contains the virtual screen panel area outside
the panel area.

This patch changes the order to call setCalibrateAffine with
dedicated parameters for Core2 to overwrite the default
unexpected matrix.