Autodetect for Guition JC2432W328C
The device is similar to 2432S028 "cheap yellow displays", but it
has a capacitive touch screen using CS820 and the backlight is on
GPIO 27 instead of GPIO 21.  The panel is ILI9341_2.

I gave it its own board ID value so application code can support
both kinds of boards, automatically adapting to the GPIO differences.